Extract stateless function initPassword
